+++ l2-frontend/src/constants.js
@@ -1,0 +1,1 @@
+export const GENDERS = ['м', 'ж'];
